IN-100 refactor: change name from ProfileLayout to LayoutWithSidebar
